Restore TX A-MPDU aggregation
During cleanup of ifdef guards for older kernels in
commit b34bfa40566f ("mac80211 I forgot the No. ... Terrible esp"),
a block of code meant for Linux kernels < v2.6.28 was accidently
left in place.

The effect was that an AMPDU_TX_OPERATIONAL message was sent down
to the firmware for an AMPDU_TX_START notification from mac80211,
in addition to triggering an assertion message in the driver.  This
may have had additional side effects.

For reasons not entirely clear to me, this rendered TX completely
unoperational.

This patch fixes that by deleting the stray code block.